pizza

+++ variabelenSpeelhaldag.py
@@ -1,14 +1,14 @@
 toegang = 7.45
 vijf_minuten = 0.37
 
-aantal = float(input('Aantal personen: '))
+aantal = int(input('Aantal personen: '))
 aantal_tijd = int(input('Aantal pesonen tijd: '))
-lengte = float(input('Tijdsduur: '))
+lengte = int(input('Tijdsduur: '))
 
 
-tijdsDuurKoste = lengte/5
-toegangKosten = aantal * toegang
-tijdskosten =  vijf_minuten * tijdsDuurKoste * aantal_tijd
-totaalKosten = tijdskosten + toegangKosten
+tijdsDuurKoste = float(lengte/5)
+toegangKosten = float(aantal * toegang)
+tijdskosten =  float(vijf_minuten * tijdsDuurKoste * aantal_tijd)
+totaalKosten = float(tijdskosten + toegangKosten)
 
 print(str(totaalKosten) + ' euro')
